+++ src/senaite/storage/content/analysisrequest.py
@@ -21,8 +21,10 @@
 from archetypes.schemaextender.interfaces import IBrowserLayerAwareExtender
 from archetypes.schemaextender.interfaces import ISchemaExtender
 from bika.lims.interfaces import IAnalysisRequest
+from Products.Archetypes.atapi import StringWidget
 from senaite.core.browser.widgets import DateTimeWidget
 from senaite.storage.archetypes.field import ExtDateTimeField
+from senaite.storage.archetypes.field import ExtStringField
 from senaite.storage.interfaces import ISenaiteStorageLayer
 from zope.component import adapter
 from zope.interface import implementer
@@ -33,6 +35,13 @@ fields = [
         "StorageExpiryDate",
         default=None,
         widget=DateTimeWidget(
+            visible=False,
+        ),
+    ),
+    ExtStringField(
+        "RetrieveReason",
+        default="",
+        widget=StringWidget(
             visible=False,
         ),
     ),